Fix float entering

+++ printLabel/views.py
@@ -51,10 +51,10 @@ def setup(request):
 
 def nutritionFacts(request):
     nutritionName = request.POST.get('nutritionName')
-    X = int(request.POST.get('nutritionX')) *8
-    Y = int(request.POST.get('nutritionY')) *8
+    X = float(request.POST.get('nutritionX')) *8
+    Y = float(request.POST.get('nutritionY')) *8
     option = request.POST.getlist('options')
-    weight = int(request.POST.get('weight'))
+    weight = float(request.POST.get('weight'))
     servings = request.POST.get('servings')
     optionList = request.POST.get('optionList')
     optionList = list(eval(optionList).keys())
